// src/app/features/networkdata/components/stops/interfaces/stoppoint.interface.ts
export interface IStopPoint {
  primarykeyduid?: number;
  stopId?:number;
  stopname?:string;
  stoppointnumber?: number;
  longdescription?: string;
  geonodeno?: number;
  dfizieltext?: number;  
  latitude?: number;
  longitude?: number;
  gpskompassrichtung?: number;
  netzpunktindex?:number;
}